+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelog
 
 ## 1.0.3
+
 - Set subdivisions slider input step to 1 in grid example.
+- Added external camera option to DualCameraControls
+- Clean up viewHelper example
 
 ## 1.0.2
+
 - View helper has a transparent background.
 - View helper default labels are x, y, and z.